create a new form for the listing form and remove some unnecessary code
