+++ frontend/src/components/sellListing/ListingForm.tsx
@@ -1,36 +1,44 @@
-import { useEffect, useState } from 'react';
+import { SelectChangeEvent, Button, TextField, Select, MenuItem } from '@mui/material';
+import { useState, useEffect } from 'react';
 import { useLoaderData, useParams, useNavigate } from 'react-router-dom';
 import {
-  ListingDetailsLoaderData,
-  YugiohCardInSet,
+  CardDetailsLoaderData,
   YugiohCardListing,
+  condition,
   YugiohCardSellListing,
 } from '../../services/yugioh/types';
-import { useCurrentUser } from '../../context/user';
 import { yugiohService } from '../../services/yugioh/yugiohService';
+import { errorToast } from '../../util/errorToast';
 import { useEffectAfterInitialLoad } from '../../util/useEffectAfterInitialLoad';
-import YugiohCardMarket from '../../components/yugioh/table/market/YugiohCardMarket';
-import CardflowTabs from '../../components/sellListing/CardflowTabs';
-import NewListingTopBar from '../../components/sellListing/NewListingTopBar';
+import YugiohCardMarket from '../yugioh/table/market/YugiohCardMarket';
+import CardflowTabs from './CardflowTabs';
+import NewListingTopBar from './NewListingTopBar';
 import PaymentsIcon from '@mui/icons-material/Payments';
+import DeleteIcon from '@mui/icons-material/Delete';
 import DiamondIcon from '@mui/icons-material/Diamond';
 import TagIcon from '@mui/icons-material/Tag';
 import VisibilityIcon from '@mui/icons-material/Visibility';
-import { errorToast } from '../../util/errorToast';
-import toast from 'react-hot-toast';
-import { toastMessages } from '../../constants/toast';
 
-function EditListing(): JSX.Element {
+const selectOptions: Record<condition, string> = {
+  poor: 'Poor',
+  played: 'Played',
+  good: 'Good',
+  excellent: 'Excellent',
+};
+
+type ListingFormProps = {
+  editMode?: boolean;
+  onSubmit: (data: YugiohCardSellListing, postAnother: boolean) => void;
+};
+
+function ListingForm(props: ListingFormProps) {
   const [page, setPage] = useState(1);
-  const data = useLoaderData() as ListingDetailsLoaderData;
+  const data = useLoaderData() as CardDetailsLoaderData;
   const params = useParams();
   const id = Number(params.id);
-  const { cardListings: cardListingsData } = data;
+  const { cardInSet, cardListings: cardListingsData } = data;
   const [cardListings, setCardListings] = useState(cardListingsData);
   const navigate = useNavigate();
-  const { user: currentUser } = useCurrentUser();
-  const [cardId, setCardId] = useState(0);
-  const [cardInSet, setCardInSet] = useState<YugiohCardInSet>();
 
   const [formData, setFormData] = useState<YugiohCardListing>({
     id: 0,
@@ -93,13 +101,7 @@ function EditListing(): JSX.Element {
           card_set_id,
         } = currentCard;
         if (currentCard && currentCard !== null) {
-          if (currentUser.user_id !== user) {
-            navigate('/sell/manage');
-          }
-          setCardId(card);
-          const cardInSet = await yugiohService.getCardInSetById(card);
-          setCardInSet(cardInSet);
-          setFormData(() => ({
+          setFormData({
             card,
             quantity,
             condition,
@@ -138,80 +140,82 @@ function EditListing(): JSX.Element {
                 rarity_code: '',
               },
             },
-          }));
+          });
         }
       } catch (error) {
-        errorToast(error);
+        errorToast(error, undefined, 404);
       }
     }
+
     loadCardListing();
-  }, [page]);
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
-    const { name, value, type } = e.target;
+  }, [params.id]);
 
-    let newValue: string | boolean = value;
+  const handleQuantityChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const value = e.target.value;
+    const quantity = Number(value);
 
-    if (type === 'checkbox') {
-      // 'checked' is only a property on input elements, specifically checkboxes
-      newValue = (e.target as HTMLInputElement).checked;
+    if (quantity > 0) {
+      setFormData((prevFormData) => ({
+        ...prevFormData,
+        quantity,
+      }));
     }
+  };
 
-    setFormData((prevFormData) => ({
-      ...prevFormData,
-      [name]: newValue,
-    }));
+  const handlePriceChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const value = e.target.value;
+    const price = Number(value);
+
+    // Allow zero so users can input prices like 0.99 without issues
+    if (price >= 0) {
+      setFormData((prevFormData) => ({
+        ...prevFormData,
+        price,
+      }));
+    }
+  };
+
+  const handleConditionChange = (e: SelectChangeEvent<condition>) => {
+    const condition = e.target.value as condition;
+    if (condition) {
+      setFormData((prevFormData) => ({
+        ...prevFormData,
+        condition,
+      }));
+    }
   };
 
   async function deleteListing() {
     try {
       await yugiohService.deleteListingById(Number(params.id));
       navigate('/sell/manage');
-      toast.success(
-        toastMessages.success.listingDeleted(
-          cardInSet?.yugioh_card.card_name || '',
-          cardInSet?.set.set_code || '',
-        ),
-      );
     } catch (error) {
       errorToast(error);
     }
-  }
-  function delistItem() {
-    yugiohService
-      .editListing({ ...formData, is_listed: !formData.is_listed })
-      .then((data) =>
-        toast.success(
-          toastMessages.success.listingVisibilityChanged(
-            data.card_name,
-            data.card_in_set.set.set_code,
-            !formData.is_listed,
-          ),
-        ),
-      )
-      .catch(errorToast);
   }
 
-  async function updateListing(e: React.FormEvent): Promise<void> {
+  function delistItem() {
+    yugiohService.editListing({ ...formData, is_listed: !formData.is_listed }).catch(errorToast);
+  }
+
+  async function handleSubmit(e: React.FormEvent, postAnother: boolean): Promise<void> {
     e.preventDefault();
-    try {
-      const newData: Partial<YugiohCardSellListing> = {
-        quantity: Number(formData.quantity),
-        condition: formData.condition,
-        price: Number(formData.price),
-        is_listed: formData.is_listed,
-      };
-      const listing = await yugiohService.updateCardListing(newData, id);
-      toast.success(
-        toastMessages.success.listingUpdated(listing.card_name, listing.card_in_set.set.set_code),
-      );
-    } catch (error) {
-      errorToast(error);
-    }
+
+    const newData: YugiohCardSellListing = {
+      quantity: formData.quantity,
+      condition: formData.condition,
+      price: formData.price,
+      is_sold: false,
+      is_listed: formData.is_listed,
+      card: id,
+    };
+
+    props.onSubmit(newData, postAnother);
   }
 
   function changePage(page: number) {
     yugiohService
-      .getCardListingsByCardSetId(cardId, page)
+      .getCardListingsByCardSetId(id, page)
       .then((data) => {
         setCardListings(data);
         setPage(page);
@@ -220,101 +224,107 @@ function EditListing(): JSX.Element {
   }
 
   useEffectAfterInitialLoad(() => {
-    yugiohService.getCardListingsByCardSetId(cardId).then(setCardListings).catch(errorToast);
+    yugiohService.getCardListingsByCardSetId(id).then(setCardListings).catch(errorToast);
     setPage(1);
-  }, [cardId]);
+  }, [id]);
   return (
     <section className="bg-[#F5F5F5]">
       <CardflowTabs />
       <NewListingTopBar
-        handleSubmit={updateListing}
+        handleSubmit={handleSubmit}
         quantity={formData.quantity}
         price={formData.price}
       />
       <div className="block mt-20 ml-40 mr-40 bg-white rounded-lg">
         <div className="flex flex-col border border-stone-300">
           <div className="bg-white ml-4 p-8 w-[314px] h-[422px] flex">
-            <img src={cardInSet?.yugioh_card.image} />
+            <img src={cardInSet.yugioh_card.image} />
             <div>
               <div>
                 <h3
                   className="ml-8 text-black text-nowrap text-4xl font-medium"
                   style={{ whiteSpace: 'nowrap' }}
                 >
-                  {cardInSet?.yugioh_card.card_name}
+                  {cardInSet.yugioh_card.card_name}
                 </h3>
                 <div className="absolute right-48 top-80">
-                  <button
-                    onClick={deleteListing}
-                    className=" flex justify-center items-center w-20 h-8  mt-4 text-sm text-gray-700 border border-stone-300 pl-3 pr-4 rounded-md hover:bg-gray-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-gray-500"
-                  >
-                    Delete
-                  </button>
-
-                  <button
-                    onClick={delistItem}
-                    className=" flex justify-center items-center w-20 h-8  mt-4 text-sm text-gray-700 border border-stone-300 pl-3 pr-4 rounded-md hover:bg-gray-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-gray-500"
-                  >
-                    <VisibilityIcon className="p-1" />
-                    Delist
-                  </button>
+                  {id ? (
+                    <Button color="error" variant="outlined" onClick={deleteListing}>
+                      Delete
+                    </Button>
+                  ) : null}
+                  {id ? null : (
+                    <Button href="/sell/new" variant="outlined" startIcon={<DeleteIcon />}>
+                      Clear
+                    </Button>
+                  )}
+                  {id ? (
+                    <Button onClick={delistItem} variant="outlined" startIcon={<VisibilityIcon />}>
+                      Delist
+                    </Button>
+                  ) : null}
                 </div>
               </div>
               <div className="flex">
                 <div className="w-[61px] h-[23px] rounded-[10px] text-neutral-500 text-sm font-normal text-center ml-8 mt-4 border border-neutral-300">
-                  {cardInSet?.set.set_code}
+                  {cardInSet.set.set_code}
                 </div>
                 <div className="w-[61px] h-[23px] rounded-[10px] text-neutral-500 text-sm font-normal text-center ml-4 mt-4 border border-neutral-300">
-                  {cardInSet?.rarity.rarity_code}
+                  {cardInSet.rarity.rarity_code}
                 </div>
               </div>
               <div className="flex mt-20">
-                <div className="inline-flex items-center p-1 ml-8 border rounded-md border-gray-300">
-                  <span className="pl-1 pr-1 text-gray-700 text-sm flex items-center">
-                    <TagIcon />
-                  </span>
-                  <input
+                <div className="w-40">
+                  <TextField
                     type="number"
                     name="quantity"
-                    min="0"
                     value={formData.quantity}
-                    onChange={handleChange}
-                    className="w-24 px-2 py-0.5 text-sm text-gray-700 border-none focus:ring-0 focus:outline-none"
+                    onChange={handleQuantityChange}
                     placeholder="Quantity"
+                    size="small"
+                    InputProps={{
+                      startAdornment: <TagIcon className="mr-2" />,
+                    }}
                   />
                 </div>
-                <div className="inline-flex items-center p-1 ml-8 border rounded-md border-gray-300">
-                  <span className="pl-1 pr-1 text-gray-700 text-sm flex items-center">
-                    <DiamondIcon />
-                  </span>
-                  <select
+                <div>
+                  <Select
                     id="condition"
                     name="condition"
+                    size="small"
+                    placeholder="Condition"
                     value={formData.condition}
-                    onChange={handleChange}
-                    className="w-28 px-2 py-0.5 text-sm text-left text-gray-700 border-none focus:ring-0 focus:outline-none"
+                    onChange={handleConditionChange}
+                    className="w-40"
+                    renderValue={(value) => (
+                      <>
+                        <DiamondIcon className="mr-2" />
+                        <span>{selectOptions[value]}</span>
+                      </>
+                    )}
                   >
-                    <option value="poor">Poor</option>
-                    <option value="played">Played</option>
-                    <option value="good">Good</option>
-                    <option value="excellent">Excellent</option>
-                  </select>
+                    <MenuItem disabled>
+                      <i>Condition</i>
+                    </MenuItem>
+                    <MenuItem value="poor">Poor</MenuItem>
+                    <MenuItem value="played">Played</MenuItem>
+                    <MenuItem value="good">Good</MenuItem>
+                    <MenuItem value="excellent">Excellent</MenuItem>
+                  </Select>
                 </div>
               </div>
-
               <div>
-                <div className="inline-flex items-center p-1 mt-16 ml-8 border rounded-md border-gray-300">
-                  <span className="pl-1 pr-1 text-gray-700 text-sm flex items-center">
-                    <PaymentsIcon />
-                  </span>
-                  <input
+                <div className="w-40">
+                  <TextField
                     type="number"
                     name="price"
                     value={formData.price}
-                    onChange={handleChange}
-                    min="0"
-                    className="w-24 px-2 py-0.5 text-sm text-gray-700 border-none focus:ring-0 focus:outline-none"
+                    onChange={handlePriceChange}
                     placeholder="Price"
+                    size="small"
+                    InputProps={{
+                      startAdornment: <PaymentsIcon className="mr-2" />,
+                    }}
                   />
                 </div>
               </div>
@@ -335,4 +345,4 @@ function EditListing(): JSX.Element {
   );
 }
 
-export default EditListing;
+export default ListingForm;
