+++ frontend/src/pages/listing/NewListing.tsx
@@ -7,10 +7,8 @@ function Newlisting() {
     <>
       <CardflowTabs />
       <NewListingTopBar
-        card={0}
         quantity={0}
         price={0}
-        condition={'poor'}
         handleSubmit={(): Promise<void> => {
           throw new Error('Function not implemented.');
         }}
